+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+### Fixed
+
+- Fix removing hidden lines of code blocks that are quoted
+
 ## [0.18.0] - 2025-08-30
 
 ### Added